+++ kodilla-hibernate/src/main/java/com/kodilla/hibernate/invoice/dao/ItemDao.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 @Transactional
 @Repository()
-public interface ItemDao extends CrudRepository <Item, Integer> {
+public interface ItemDao extends CrudRepository <Item, Long> {
 
 }
